fix(review): isolate GitHub review mutation coordination
