+++ src/autoskillit/execution/github_review/__init__.py
@@ -1,6 +1,6 @@
 """Authoritative GitHub pull-request review publication."""
 
-from ._poster_support import GitHubReviewMutationCoordinator
+from ._mutation_coordinator import GitHubReviewMutationCoordinator
 from .canonical import canonicalize_review_request, compute_review_operation_key
 from .gateway import DefaultGitHubReviewGateway
 from .ledger import GitHubReviewLedger
